requirements.txt
langchain
langchain_ollama
sqlite3
python-dotenv
matplotlib
seaborn
streamlit
requests
pandas

